poodle/backend/requirements.txt
blinker==1.6.2
click==8.1.3
cryptography==3.4.8
distro==1.7.0
Flask==2.3.2
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
more-itertools==8.10.0
oauthlib==3.2.0
packaging==23.1
PyJWT==2.3.0
pyparsing==2.4.7
SecretStorage==3.3.1
six==1.16.0
typing_extensions==4.6.3
wadllib==1.3.6
Werkzeug==2.3.6
zipp==1.0.0
